feat: Enhance morph target processing in WebGL shaders
- Updated vertex shader functions to incorporate blend shape component IDs for improved morph target handling.
- Refactored position calculation logic to utilize uniform buffers for morph offsets and weights, enhancing animation fidelity.
- Added new uniform for current primitive index to facilitate dynamic morph target updates based on the active primitive.

+++ src/webgl/shaderity_shaders/common/prerequisites.glsl
@@ -13,7 +13,7 @@ uniform highp sampler2D u_dataTexture; // skipProcess=true
 #endif
 
 layout (std140) uniform UniformMorphOffsets {
-  vec4 data[/* shaderity: @{maxMorphDataNumber} */];
+  ivec4 data[/* shaderity: @{maxMorphDataNumber} */];
 } uniformMorphOffsets;
 
 layout (std140) uniform UniformMorphWeights {
